Decompose AgentSession and desktop store helpers
Extract AgentSession runtime/snapshot support services and move desktop feed/transcript mapping logic into a dedicated module while preserving public APIs.
